recommender-base/requirements.txt
flask
flask-sqlalchemy
flask-user==1.0.2.2
chromadb
